+++ jdaviz/configs/imviz/plugins/aper_phot_simple/aper_phot_simple.vue
@@ -1,0 +1,76 @@
+<template>
+  <v-card flat tile>
+    <v-card>
+      <v-card-text>
+        <v-container>
+          <j-docs-link :link="'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#aper-phot-simple'">
+            Perform aperture photometry for a single region. See the <j-external-link link="https://photutils.readthedocs.io/en/stable/aperture.html" linktext="photutils docs"/> for more details.
+          </j-docs-link>
+          <v-row>
+            <v-col>
+              <v-select
+                :items="viewer_items"
+                @change="viewer_selected"
+                label="Viewer"
+                hint="Select viewer for photometry."
+                persistent-hint
+              ></v-select>
+            </v-col>
+          </v-row>
+          <v-row>
+            <v-col>
+              <v-select
+                :items="dc_items"
+                @change="data_selected"
+                label="Data"
+                hint="Select data for photometry."
+                persistent-hint
+              ></v-select>
+            </v-col>
+          </v-row>
+          <v-row>
+            <v-col>
+              <v-select
+                :items="subset_items"
+                @change="subset_selected"
+                label="Subset"
+                hint="Select subset region for photometry."
+                persistent-hint
+              ></v-select>
+            </v-col>
+          </v-row>
+        </v-container>
+      </v-card-text>
+      <v-divider></v-divider>
+
+      <v-card-text v-if="result_available">
+        <v-container>
+          <v-row>
+            <v-col cols=6><U>Function</U></v-col>
+            <v-col cols=6><U>Result</U></v-col>
+          </v-row>
+          <v-row
+            v-for="item in results"
+            :key="item.function">
+            <v-col cols=6>
+              {{  item.function  }}
+            </v-col>
+            <v-col cols=6>{{ item.result }}</v-col>
+          </v-row>
+        </v-container>
+      </v-card-text>
+      <v-divider></v-divider>
+
+      <v-card-actions>
+        <div class="flex-grow-1"></div>
+        <v-row
+          justify="left"
+          align="center"
+          no-gutters
+        >
+          <v-btn color="primary" text @click="do_aper_phot">Go</v-btn>
+        </v-row>
+      </v-card-actions>
+    </v-card>
+  </v-card>
+</template>